Fixing a bug in the bitblt and testing inlining
the bitblt function in the win32 platform source and dest was flipped.
Also, I want to inline the clock but I first want to make sure that the
compiler will inline  it before i change thigns to pass by value. I
messed with the disassembler a bit and figured out how to detect inlined
vs no inlined functions (call process). tomorrow I will try to inline
the clock and see if the compiler likes that.