Inline test 2
So I did some tests to clock and I didn't really like what I found:
testfunc(), testfunc2() I could not inline. If i did a cast from
Igame_clock to win32_clock and called testfunc3() it inlined perfectly.
I found this which makes alot of sense:

When the object is referenced via a pointer or a reference, a call to a
virtual function cannot be inlined, since the call must be resolved
dynamically. Reason: the compiler can't know which actual code to call
until run-time (i.e., dynamically), since the code may be from a derived
class that was created after the caller was compiled.

Therefore the only time an inline virtual call can be inlined is when
the compiler knows the "exact class" of the object which is the target
of the virtual function call. This can happen only when the compiler has
an actual object rather than a pointer or reference to an object. I.e.,
either with a local object, a global/static object, or a fully contained
object inside a composite.

I may have to rethink a few things because of this.