Inline test 3 + cutting out fat
So i finished up my  inline tests. turns out it is alot harder to inline
than i though. for instance, the compiler rejects:

void win32_platform::paintWindow(HDC deviceContext)
{
// Blt the dib section with the option to scale
// Note: this is bitblt with scaling
StretchDIBits(deviceContext,
0, 0, _clientWidth, _clientHeight,    // dest size
0, 0, _backBuffer.width, _backBuffer.height, // source size
_backBuffer.memory,
&_bitmapInfo,
DIB_RGB_COLORS,   // DIB section will not use a palette
SRCCOPY);    // copy the src to the dest
}

to inline. I have no idea why. but I refactored how some if the stuff
works. removed the stickman_engine class which was just bloat for the
sake of a bit of cleaner code but it was causing round trips from the
platform to the game_engine. so now all of that stuff is just done in
the platform and there is no need for a platform interface.

TODO: handle the clock. I probably just want some memory and have the
platform handle it.

TODO: Get rid of Igame_io.h. Need something better for accessing
platform io stuff from the game.