Removing more oo code. Getting the project back to a buildable state.
Got rid of the game IO interfaces because we wont be able to inline
using it. Moved callback to common.

TODO: Figure out how to bind the function pointers (they currently still
do not work). I would prefer to not have to use the callback class when
it will have alot more overhead.